+++ app/Http/Livewire/Admin/GuiasRemision/Edit.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace App\Http\Livewire\Admin\GuiasRemision;
+
+use Livewire\Component;
+
+class Edit extends Component
+{
+    public function render()
+    {
+        return view('livewire.admin.guias-remision.edit');
+    }
+}
