+++ resources/views/admin/almacen/dispositivos/modelos-index.blade.php
@@ -2,51 +2,48 @@
 @section('ruta', 'almacen-dispositivos')
 @section('contenido')
 
-<!-- Table -->
-@livewire('admin.dispositivos.modelos-dispositivos-index')
+    <!-- Table -->
+    @livewire('admin.dispositivos.modelos-dispositivos-index')
 
 @stop
 
 @push('modals')
-@livewire('admin.dispositivos.edit-modelo-dispositivo')
+    @livewire('admin.dispositivos.edit-modelo-dispositivo')
 @endpush
 
 @section('js')
-<script>
-    console.log('Hi!'); 
 
-</script>
 
-<script>
-    // A basic demo function to handle "select all" functionality
-    document.addEventListener('alpine:init', () => {
-        Alpine.data('handleSelect', () => ({
-            selectall: false,
-            selectAction() {
-                countEl = document.querySelector('.table-items-action');
-                if (!countEl) return;
-                checkboxes = document.querySelectorAll('input.table-item:checked');
-                document.querySelector('.table-items-count').innerHTML = checkboxes.length;
-                if (checkboxes.length > 0) {
-                    countEl.classList.remove('hidden');
-                } else {
-                    countEl.classList.add('hidden');
+    <script>
+        // A basic demo function to handle "select all" functionality
+        document.addEventListener('alpine:init', () => {
+            Alpine.data('handleSelect', () => ({
+                selectall: false,
+                selectAction() {
+                    countEl = document.querySelector('.table-items-action');
+                    if (!countEl) return;
+                    checkboxes = document.querySelectorAll('input.table-item:checked');
+                    document.querySelector('.table-items-count').innerHTML = checkboxes.length;
+                    if (checkboxes.length > 0) {
+                        countEl.classList.remove('hidden');
+                    } else {
+                        countEl.classList.add('hidden');
+                    }
+                },
+                toggleAll() {
+                    this.selectall = !this.selectall;
+                    checkboxes = document.querySelectorAll('input.table-item');
+                    [...checkboxes].map((el) => {
+                        el.checked = this.selectall;
+                    });
+                    this.selectAction();
+                },
+                uncheckParent() {
+                    this.selectall = false;
+                    document.getElementById('parent-checkbox').checked = false;
+                    this.selectAction();
                 }
-            },
-            toggleAll() {
-                this.selectall = !this.selectall;
-                checkboxes = document.querySelectorAll('input.table-item');
-                [...checkboxes].map((el) => {
-                    el.checked = this.selectall;
-                });
-                this.selectAction();
-            },
-            uncheckParent() {
-                this.selectall = false;
-                document.getElementById('parent-checkbox').checked = false;
-                this.selectAction();
-            }
-        }))
-    })    
-</script>
+            }))
+        })
+    </script>
 @stop
