+++ resources/views/livewire/admin/guias-remision/edit.blade.php
@@ -1,0 +1,3 @@
+<div>
+    {{-- The best athlete wants his opponent at his best. --}}
+</div>
